Splitting config parser
