+++ src/internal/config/parser/cache/mod.rs
@@ -1,0 +1,8 @@
+mod root;
+pub(crate) use root::CacheConfig;
+
+mod asdf;
+pub(crate) use asdf::AsdfCacheConfig;
+
+mod homebrew;
+pub(crate) use homebrew::HomebrewCacheConfig;
